Split serde mod into muliple modules
