+++ lib/src/types/serde/deser.rs
@@ -1,3 +1,4 @@
+use super::DeError;
 use crate::types::{
     BoltInteger, BoltMap, BoltNode, BoltRelation, BoltString, BoltType, BoltUnboundedRelation,
 };
@@ -5,270 +6,11 @@ use crate::types::{
 use serde::{
     de::{
         value::{BorrowedStrDeserializer, I64Deserializer, MapDeserializer, SeqDeserializer},
-        Deserializer, Error, Expected, IntoDeserializer, Unexpected as Unexp, Visitor,
+        Deserializer, Error, IntoDeserializer, Unexpected as Unexp, Visitor,
     },
-    forward_to_deserialize_any, Deserialize,
+    forward_to_deserialize_any,
 };
-use std::{collections::HashSet, fmt, iter, marker::PhantomData, sync::Arc};
-
-/// Newtype to extract the node id or relationship id during deserialization.
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Hash, Deserialize)]
-pub struct Id(pub u64);
-
-/// Newtype to extract the start node id of a relationship during deserialization.
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Hash, Deserialize)]
-pub struct StartNodeId(pub u64);
-
-/// Newtype to extract the end node id of a relationship during deserialization.
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Hash, Deserialize)]
-pub struct EndNodeId(pub u64);
-
-/// Newtype to extract the node labels during deserialization.
-#[derive(Clone, Debug, Default, PartialEq, Eq, Hash, Deserialize)]
-pub struct Labels<Coll = Vec<String>>(pub Coll);
-
-/// Newtype to extract the relationship type during deserialization.
-#[derive(Clone, Debug, Default, PartialEq, Eq, Hash, Deserialize)]
-pub struct Type<T = String>(pub T);
-
-/// Newtype to extract the node property keys during deserialization.
-#[derive(Clone, Debug, Default, PartialEq, Eq, Deserialize)]
-pub struct Keys<Coll = HashSet<String>>(pub Coll);
-
-impl BoltMap {
-    pub(crate) fn to<'this, T>(&'this self) -> Result<T, DeError>
-    where
-        T: Deserialize<'this>,
-    {
-        T::deserialize(MapDeserializer::new(self.value.iter()))
-    }
-}
-
-impl BoltNode {
-    pub(crate) fn to<'this, T>(&'this self) -> Result<T, DeError>
-    where
-        T: Deserialize<'this>,
-    {
-        T::deserialize(self.into_deserializer())
-    }
-}
-
-impl BoltRelation {
-    pub(crate) fn to<'this, T>(&'this self) -> Result<T, DeError>
-    where
-        T: Deserialize<'this>,
-    {
-        T::deserialize(self.into_deserializer())
-    }
-}
-
-impl BoltUnboundedRelation {
-    pub(crate) fn to<'this, T>(&'this self) -> Result<T, DeError>
-    where
-        T: Deserialize<'this>,
-    {
-        T::deserialize(self.into_deserializer())
-    }
-}
-
-#[cfg(test)]
-impl BoltType {
-    pub(crate) fn to<'this, T>(&'this self) -> Result<T, DeError>
-    where
-        T: Deserialize<'this>,
-    {
-        T::deserialize(self.into_deserializer())
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum DeError {
-    /// Raised when a `Deserialize` receives a type different from what it was
-    /// expecting.
-    #[error("Invalid type: {received}, expected {expected}")]
-    InvalidType {
-        received: Unexpected,
-        expected: String,
-    },
-
-    /// Raised when a `Deserialize` receives a value of the right type but that
-    /// is wrong for some other reason.
-    #[error("Invalid value: {received}, expected {expected}")]
-    InvalidValue {
-        received: Unexpected,
-        expected: String,
-    },
-
-    /// Raised when deserializing a sequence or map and the input data contains
-    /// too many or too few elements.
-    #[error("Invalid length {received}, expected {expected}")]
-    InvalidLength { received: usize, expected: String },
-
-    /// Raised when a `Deserialize` enum type received a variant with an
-    /// unrecognized name.
-    #[error("Unknown variant `{variant}`, expected {expected:?}")]
-    UnknownVariant {
-        variant: String,
-        expected: &'static [&'static str],
-    },
-
-    /// Raised when a `Deserialize` struct type received a field with an
-    /// unrecognized name.
-    #[error("Unknown field `{field}`, expected {expected:?}")]
-    UnknownField {
-        field: String,
-        expected: &'static [&'static str],
-    },
-
-    /// Raised when a `Deserialize` struct type expected to receive a required
-    /// field with a particular name but that field was not present in the
-    /// input.
-    #[error("Missing field `{field}`")]
-    MissingField { field: &'static str },
-
-    /// Raised when a `Deserialize` struct type received more than one of the
-    /// same field.
-    #[error("Duplicate field `{field}`")]
-    DuplicateField { field: &'static str },
-
-    #[error("The property `{0}` does not exist")]
-    NoSuchProperty(String),
-
-    #[error("{0}")]
-    Other(String),
-
-    #[error("Could not convert the integer `{1}` to the target type {2}")]
-    IntegerOutOfBounds(#[source] std::num::TryFromIntError, i64, &'static str),
-}
-
-/// `Unexpected` represents an unexpected invocation of any one of the `Visitor`
-/// trait methods.
-///
-/// This mirrors the [`serde::de::Unexpected`] type, but uses owned types
-/// instead of borrowed types.
-///
-/// The owned typed in question are `String` and `Vec<u8>`, which are stored
-/// as `Arc<str>` and `Arc<[u8]>` respectively, so that this type is cheap-ish
-/// to clone and can be shared between threads.
-#[derive(Clone, PartialEq, Debug)]
-pub enum Unexpected {
-    /// The input contained a boolean value that was not expected.
-    Bool(bool),
-
-    /// The input contained an unsigned integer `u8`, `u16`, `u32` or `u64` that
-    /// was not expected.
-    Unsigned(u64),
-
-    /// The input contained a signed integer `i8`, `i16`, `i32` or `i64` that
-    /// was not expected.
-    Signed(i64),
-
-    /// The input contained a floating point `f32` or `f64` that was not
-    /// expected.
-    Float(f64),
-
-    /// The input contained a `char` that was not expected.
-    Char(char),
-
-    /// The input contained a `&str` or `String` that was not expected.
-    Str(Arc<str>),
-
-    /// The input contained a `&[u8]` or `Vec<u8>` that was not expected.
-    Bytes(Arc<[u8]>),
-
-    /// The input contained a unit `()` that was not expected.
-    Unit,
-
-    /// The input contained an `Option<T>` that was not expected.
-    Option,
-
-    /// The input contained a newtype struct that was not expected.
-    NewtypeStruct,
-
-    /// The input contained a sequence that was not expected.
-    Seq,
-
-    /// The input contained a map that was not expected.
-    Map,
-
-    /// The input contained an enum that was not expected.
-    Enum,
-
-    /// The input contained a unit variant that was not expected.
-    UnitVariant,
-
-    /// The input contained a newtype variant that was not expected.
-    NewtypeVariant,
-
-    /// The input contained a tuple variant that was not expected.
-    TupleVariant,
-
-    /// The input contained a struct variant that was not expected.
-    StructVariant,
-
-    /// A message stating what uncategorized thing the input contained that was
-    /// not expected.
-    ///
-    /// The message should be a noun or noun phrase, not capitalized and without
-    /// a period. An example message is "unoriginal superhero".
-    Other(Arc<str>),
-}
-
-impl From<Unexp<'_>> for Unexpected {
-    fn from(value: Unexp<'_>) -> Self {
-        match value {
-            Unexp::Bool(v) => Self::Bool(v),
-            Unexp::Unsigned(v) => Self::Unsigned(v),
-            Unexp::Signed(v) => Self::Signed(v),
-            Unexp::Float(v) => Self::Float(v),
-            Unexp::Char(v) => Self::Char(v),
-            Unexp::Str(v) => Self::Str(v.into()),
-            Unexp::Bytes(v) => Self::Bytes(v.into()),
-            Unexp::Unit => Self::Unit,
-            Unexp::Option => Self::Option,
-            Unexp::NewtypeStruct => Self::NewtypeStruct,
-            Unexp::Seq => Self::Seq,
-            Unexp::Map => Self::Map,
-            Unexp::Enum => Self::Enum,
-            Unexp::UnitVariant => Self::UnitVariant,
-            Unexp::NewtypeVariant => Self::NewtypeVariant,
-            Unexp::TupleVariant => Self::TupleVariant,
-            Unexp::StructVariant => Self::StructVariant,
-            Unexp::Other(v) => Self::Other(v.into()),
-        }
-    }
-}
-
-impl Unexpected {
-    fn into(&self) -> Unexp<'_> {
-        match self {
-            Unexpected::Bool(v) => Unexp::Bool(*v),
-            Unexpected::Unsigned(v) => Unexp::Unsigned(*v),
-            Unexpected::Signed(v) => Unexp::Signed(*v),
-            Unexpected::Float(v) => Unexp::Float(*v),
-            Unexpected::Char(v) => Unexp::Char(*v),
-            Unexpected::Str(v) => Unexp::Str(v),
-            Unexpected::Bytes(v) => Unexp::Bytes(v),
-            Unexpected::Unit => Unexp::Unit,
-            Unexpected::Option => Unexp::Option,
-            Unexpected::NewtypeStruct => Unexp::NewtypeStruct,
-            Unexpected::Seq => Unexp::Seq,
-            Unexpected::Map => Unexp::Map,
-            Unexpected::Enum => Unexp::Enum,
-            Unexpected::UnitVariant => Unexp::UnitVariant,
-            Unexpected::NewtypeVariant => Unexp::NewtypeVariant,
-            Unexpected::TupleVariant => Unexp::TupleVariant,
-            Unexpected::StructVariant => Unexp::StructVariant,
-            Unexpected::Other(v) => Unexp::Other(v),
-        }
-    }
-}
-
-impl fmt::Display for Unexpected {
-    fn fmt(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
-        self.into().fmt(formatter)
-    }
-}
+use std::{iter, marker::PhantomData, result::Result};
 
 pub struct BoltTypeDeserializer<'de> {
     value: BoltRef<'de>,
@@ -1005,58 +747,6 @@ impl<'de> AdditionalData<'de> for &'de BoltUnboundedRelation {
     }
 }
 
-impl<'de> Error for DeError {
-    fn invalid_type(unexp: Unexp, exp: &dyn Expected) -> Self {
-        Self::InvalidType {
-            received: unexp.into(),
-            expected: exp.to_string(),
-        }
-    }
-
-    fn invalid_value(unexp: Unexp, exp: &dyn Expected) -> Self {
-        Self::InvalidValue {
-            received: unexp.into(),
-            expected: exp.to_string(),
-        }
-    }
-
-    fn invalid_length(len: usize, exp: &dyn Expected) -> Self {
-        Self::InvalidLength {
-            received: len,
-            expected: exp.to_string(),
-        }
-    }
-
-    fn unknown_variant(variant: &str, expected: &'static [&'static str]) -> Self {
-        Self::UnknownVariant {
-            variant: variant.to_string(),
-            expected,
-        }
-    }
-
-    fn unknown_field(field: &str, expected: &'static [&'static str]) -> Self {
-        Self::UnknownField {
-            field: field.to_string(),
-            expected,
-        }
-    }
-
-    fn missing_field(field: &'static str) -> Self {
-        Self::MissingField { field }
-    }
-
-    fn duplicate_field(field: &'static str) -> Self {
-        Self::DuplicateField { field }
-    }
-
-    fn custom<T>(msg: T) -> Self
-    where
-        T: fmt::Display,
-    {
-        Self::Other(msg.to_string())
-    }
-}
-
 impl<'de> IntoDeserializer<'de, DeError> for &'de BoltType {
     type Deserializer = BoltTypeDeserializer<'de>;
 
@@ -1469,8 +1159,9 @@ mod tests {
         sync::atomic::{AtomicU32, Ordering},
     };
 
-    use super::*;
-    use crate::types::BoltNull;
+    use super::super::*;
+    use crate::types::{BoltInteger, BoltNull, BoltString};
+    use crate::{EndNodeId, Id, Keys, Labels, StartNodeId, Type};
 
     #[test]
     fn map_with_extra_fields() {
